Login feature (#2)
* Update the project folder structure.

The following folder structure is adopted.

foundations: This folder contains the core module.

features: This folder contains the different feature modules

* Create Banner ui component.

A banner UI is available to be consumed. It can show an optional icon at
the start followed by some text.

* Add the login feature.

We provide a Google Login button to the users. The final validation happens
 on the server.