[other] Refactor classes.
We now move away from inheritance style (BaseViewModel) to a composition
style (StateContainer).
The container also has support for saving UI state in a SavedStateHandle
(SavedStateContainer). We have add the Serializable annotation to our
domain entities so they can be saved in the SavedStateContainer.

We also improve the BaseUseCase and the Cancellation exception handling.